+++ src/core/index.ts
@@ -1,1 +1,5 @@
 export * from "./fastypest";
+export { ChangeDetectionStrategy } from "./types";
+export type { FastypestOptions } from "./types";
+export { LogLevel, LoggingDetailLevel } from "../logging";
+export type { LoggingOptions } from "../logging";